requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
docker==7.1.0
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
inference-sdk==0.51.1
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-ml-py==12.575.51
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
py-cpuinfo==9.0.0
pybase64==1.0.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
supervision==0.25.1
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
